Move TrafficControl class and TrafficDirection class to tcconfig/traffic_control.py
